# docs/api.yaml
openapi: 3.0.3
info:
  title: Attendify API
  version: 1.0.0
  description: |
    Attendify is a secure, distributed, cloud-native SaaS backend system.

    The platform provides:

      - Company identity management
      - Secure multi-tenant authentication
      - Edge-based request routing through Cloudflare Worker
      - JWT-based stateless authentication
      - Cryptographic request verification using HMAC-SHA256
      - Nonce-based replay protection
      - Secure attendance evidence submission

    Security model:

      - Clients communicate with the Cloudflare Worker gateway.
      - The Cloudflare Worker forwards requests to the Attendify backend.
      - The backend validates trusted gateway traffic using an internal edge secret.
      - Protected company routes require JWT Bearer authentication.
      - Critical attendance requests require nonce freshness and HMAC signature verification.

    Request flow:

      Client -> Cloudflare Worker -> Attendify Backend -> MongoDB Atlas

    Important production rule:

      Clients should use the Cloudflare Worker URL as the public API entry point.
      The Railway backend URL should not be used directly by clients.

  license:
    name: Private / Internal Use
servers:
  - url: http://localhost:3000
    description: Local Development Server

  - url: https://attendify-worker.mahmouddamour-3aa.workers.dev
    description: Production Edge Gateway
...
paths:
...
  /company/profile:
    get:
      tags:
        - Company
      operationId: getCompanyProfile
      summary: Get company profile from JWT
      description: |
        Returns lightweight company identity extracted directly from the verified JWT token.

        This endpoint is optimized for fast identity checks and does not require a database lookup.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
              example:
                success: true
                company:
                  id: 4d370664-dc1d-4b3a-9d42-ad30b67053b6
                  email: admin@acme.com

        "401":
          description: Missing, expired, or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthorized

        "403":
          description: Invalid gateway or direct backend access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/me:
    get:
      tags:
        - Company
      operationId: getFullCompanyData
      summary: Get full company data
      description: |
        Returns the full company document from the database.

        Sensitive fields such as password and apiKey must not be returned in the response.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Full company data returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResponse"
              example:
                success: true
                company:
                  id: 4d370664-dc1d-4b3a-9d42-ad30b67053b6
                  name: acme
                  email: admin@acme.com
                  createdAt: "2026-05-13T09:00:00.000Z"
                  status: active
                  loginAttempts: 0
                  lockUntil: null

        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/update:
    put:
      tags:
        - Company
      operationId: updateCompany
      summary: Update company data
      description: |
        Updates allowed company fields only.

        Allowed fields:

          - name
          - email

        Protected fields that must never be updated through this endpoint:

          - id
          - password
          - apiKey
          - status
          - loginAttempts
          - lockUntil
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdateRequest"
            example:
              name: acme-updated
              email: updated@acme.com
      responses:
        "200":
          description: Company updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
              example:
                success: true
                message: Company updated

        "400":
          description: No valid fields provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: No valid fields provided

        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/delete:
    delete:
      tags:
        - Company
      operationId: softDeleteCompany
      summary: Soft delete company
      description: |
        Performs logical deletion by marking the company status as deleted.

        The record is preserved for:

          - Auditability
          - Recovery
          - Historical tracking
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company soft deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
              example:
                success: true
                message: Company deleted soft

        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /company/lookup/{name}:
    get:
      tags:
        - Company
      operationId: lookupCompanyByName
      summary: Lookup company by name
      description: |
        Public endpoint used by clients to verify whether a company exists.

        This endpoint returns minimal safe metadata only.
        It must not return sensitive fields such as email, password, apiKey, or internal security metadata.
      parameters:
        - name: name
          in: path
          required: true
          description: Company name to search for
          schema:
            type: string
            minLength: 1
          example: acme
      responses:
        "200":
          description: Lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyLookupResponse"
              examples:
                found:
                  summary: Company exists
                  value:
                    exists: true
                    company:
                      id: 4d370664-dc1d-4b3a-9d42-ad30b67053b6
                      name: acme
                notFound:
                  summary: Company does not exist
                  value:
                    exists: false

        "400":
          description: Invalid company name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer authentication.

        Usage:

          Authorization: Bearer <token>
  schemas:
...
    SuccessMessage:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Internal server error
...
    ProfileResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: true
        company:
          $ref: "#/components/schemas/JwtCompanyIdentity"
    JwtCompanyIdentity:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: 4d370664-dc1d-4b3a-9d42-ad30b67053b6
        email:
          type: string
          format: email
          example: admin@acme.com
    CompanyResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
          example: true
        company:
          $ref: "#/components/schemas/Company"
    Company:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: 4d370664-dc1d-4b3a-9d42-ad30b67053b6
        name:
          type: string
          example: acme
        email:
          type: string
          format: email
          example: admin@acme.com
        createdAt:
          type: string
          format: date-time
          example: "2026-05-13T09:00:00.000Z"
        status:
          type: string
          example: active
        loginAttempts:
          type: integer
          example: 0
        lockUntil:
          nullable: true
          oneOf:
            - type: number
            - type: "null"
          example: null
    CompanyUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          example: acme-updated
        email:
          type: string
          format: email
          example: updated@acme.com
    CompanyLookupResponse:
      type: object
      additionalProperties: false
      properties:
        exists:
          type: boolean
          example: true
        company:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/PublicCompany"
            - type: "null"
    PublicCompany:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: 4d370664-dc1d-4b3a-9d42-ad30b67053b6
        name:
          type: string
          example: acme
...